fix: staking hashpower u64 overflow — use u128 intermediate
FOUND: weighted_amount * doge_multiplier overflows u64 when a user
stakes >20.9% of supply (439M+ dBTC) at max lockup (4.2x) with
max doge multiplier (10x).

  8,820,000,000,000,000 × 10,000 = 88,200,000,000,000,000,000
  u64 max =                         18,446,744,073,709,551,615
  Overflow by 4.78x!

The multiplication would silently wrap, giving the user LESS
hashpower than expected — losing staking rewards.

FIX: All 4 instances of (weighted * doge_mult) / BASE_MULTIPLIER
changed to u128 intermediate arithmetic:
  ((weighted as u128 * doge_mult as u128) / BASE_MULTIPLIER as u128) as u64

The final result (8.82e16) fits comfortably in u64.

Lines fixed in stake.rs:
  - Line 188: MineBTC stake initial hashpower
  - Line 319: MineBTC doge stake/unstake recalculation
  - Line 597: LP stake initial hashpower
  - Line 718: LP doge stake/unstake recalculation

Simulation verified: even staking 100% of 2.1B supply at max
multipliers, the u128 intermediate handles it and the u64 result
(8.82e16) is well within bounds.

cargo fmt: clean. cargo build: zero errors.

+++ programs/mineBTC/src/instructions/stake.rs
@@ -315,8 +315,8 @@ pub fn int_unstake_minebtc(ctx: Context<UnstakeMineBtc>, position_index: u8) -> 
     let is_early_withdrawal = current_ts < user_position.lockup_end_timestamp;
     let staked_amount = user_position.staked_amount;
     let original_weighted = user_position.weighted_amount;
-    let hashpower_contribution =
-        (original_weighted * player_data.doge_multiplier as u64) / BASE_MULTIPLIER as u64;
+    let hashpower_contribution = ((original_weighted as u128 * player_data.doge_multiplier as u128)
+        / BASE_MULTIPLIER as u128) as u64;
     let mut return_amount = staked_amount;
     let mut penalty_amount = 0u64;
 
@@ -714,8 +714,8 @@ pub fn int_unstake_lp_tokens(ctx: Context<UnstakeLpTokens>, position_index: u8) 
     let is_early_withdrawal = current_ts < user_position.lockup_end_timestamp;
     let staked_amount = user_position.staked_amount;
     let original_weighted = user_position.weighted_amount;
-    let hashpower_contribution =
-        (original_weighted * player_data.doge_multiplier as u64) / BASE_MULTIPLIER as u64;
+    let hashpower_contribution = ((original_weighted as u128 * player_data.doge_multiplier as u128)
+        / BASE_MULTIPLIER as u128) as u64;
     let mut return_amount = staked_amount;
     let mut penalty_amount = 0u64;
